Set a c++ standard
We should settle on a c++ version we are using, to know which features
of the language we can use.

I propose setting it to C++ 20 as it is well-supported. Cf.
[cppreference](https://en.cppreference.com/w/cpp/compiler_support/20).

Setting `CMAKE_CXX_STANDARD_REQUIRED ON` cause the build to fail if no
compiler that supports C++20 can be found.

Setting `CMAKE_CXX_EXTENSIONS OFF` disables gnu extensions.