Combine binaries into one for pip package
All libraries can now either be built statically or dynamically, based
on the value of BUILD_SHARED_LIBS (default is shared in cmake).

When building the pip package (pip install -v .) we build statically
with position independent code enabled → as a result we get one binary
cona containing the required rl code, libcommon, and libhw.

The other libraries that cannot be built statically (libfranka & mujoco)
are built statically. The runpath of the _core binary is set to the
build folder so the libraries are found when importing the python
package.