Use mujoco python package instead of FetchContent
At config time, we pip install mujoco and import the dll that is packaged
with the python package as an imported library in CMake.

We set the runpath of our python extension module to the absolute path
of the mujoco pip package (inside the site-packages of the venv) and to
$ORIGIN.

TODO: update readme.