Move mujoco dependency version to cmake
When building with cmake not with pip cmake needs to know what mujoco
version is being used so we can link against the correct file. Since
that was defined in pyproject.toml, the variable MUJOCO_VERSION is not
set when building with cmake --build build resulting in an error.

Solution: we move the definition of MUJOCO_VERSION to cmake.