fix(core arch): removed set_parameters in base robot class
Correct polymorphism does not a derived type in a dervived function signature which is why we used the base config type in set_parameters. But since we expected a derived type we had to cast in the derived function body. This is bad behaviour as the function caller needs to know to pass a derived type instead of the function signuture type.
The function caller has to know the type anyway to pass the correct config, thus it does not make sense to have this method at the base type and rather at the derived types. Disadvantage of this is, however, that the interface does not enforce the creation of such a set_parameters method.