start of sim refactoring
The sim is supposed to model the world. It is a separate class than can
be stepped in different ways. The different components (robots, cameras,
hands) etc. attach to a specific sim.

Components depend on the sim, the sim does not depend on any robot or
camera class. However we do not necessarily want to make every
multithreaded, that would just make things less performant since we
might want some tight control loops. If we need to compute some control
value every single step, we need to interrupt the simulation every
single step. Therefore we let different components add callbacks that
are called at a desired frequency.