Any and all callbacks instead for the sim
You can now register two condition callbacks in the sim:
an "any" callback and a "all" callback.
Both influence the Sim::step_until_convergence method.
The method will return if all the "all" callbacks return true, or if any
of the "any" callbacks return true.

Imagine a sim with two arms. We would set an action on each arm and want
the following behavior:
if both arms are done with their action we return, but, if any of the
arm collides with anything on the way, then we return early.

I.e. two callbacks: a collision callback is an any callback, a
convergence callback is a all callback.