Fix finding mujoco headers once and for all
We do it the proper way now using a FindMuJoCo.cmake file.

Adding the mujoco version to the cmake command is no longer needed.

mujoco is now a pip build dependency (we need to link against it during
build time).

This is consistent with the way other libraries do it: scikit-learn has
numpy as a build dependency for example.

Additionally no more hacky pip install in the cmake file.

During build time, pip creates a temporary venv.
In this venv mujoco is installed, then  cmake is invoked.

cmake builds the extension, links against the mujoco library in the
temporary venv.

When installing, we change the path to the mujoco library to make it
relative.