Build project with pip and without build isolation
Currently, pip creates a separate venv, installs the build dependencies
and then configures with cmake.

This results in wrong paths in the cmake cache. It breaks for example
clangd, which then looks in /tmp/pip.../mujoco/include for the mujoco
headers.

The solution is to disable build isolation. The drawback is that we have
to ensure ourselves that the build dependencies are installed.