Simplify FindMuJoCo.cmake
Before we used python to import mujoco and find where the init file of
the mujoco module is located. We did so because FindPython was reliably
detecting the python interpreter (in the venv) but not the python
sitelib directory.

This issue is fixed now that we use --no-build-isolation with pip. There
we remove the execute_process command and use the ${Python3_SITELIB}
variable.