fix: sim gripper collision in state
added gripper collision to sim franka hand state and to info dict in gripper env

This is a hot fix to allow the collision tests to pass. In the future we should follow the implementation specific wrapper approach from FR3, where sim and hw specific functionalities are put into its own wrapper.