refactor(ik): moved rl ik into common, removed rl deps in hw
This commit moves the ik code into an own class in common. This allows us to easily replace the ik without touching much code and avoid code duplication as the ik is currently  implemented in both sim and hardware. Since it defines an IK interface, it can also be later refactored into an IK base class to allow several different IK implementations.
So far this is only done for the hardware module and the sim module has for the time being only dummy methods to follow the interface.