fix: dot env in cli, panic activates guiding mode
Part of #130. Using dot env files to authenticate in desk in the desk cli.
Currently the config cannot be deleted as it is still used in the test cli for the realsense camera and the urdf path of the record cli.

Improved cli: allow guiding mode to lock and unlock, allow to use guiding mode after fci context manager finished

Removed unused guiding mode context manager

Improved documentation for examples

Fixed example imports