refactor: Remove private scenes, add public assets (#182)
* Remove private scenes, add public assets

Assets than can be shared are added in the assets dir. 

UTN scenes integration with the token etc. is removed.
Included scenes are compiled and the binary .mjb file is added to the
pip package.

The global rcsss.scenes dict now contains, for each scene, the paths to
the .mjb and the path to the .urdf.

* Removed model download from build

* Build _core and franka only in makefile